update client usage in example-code
