+++ extra/gen_rf_survey.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from shepherd_core import TestbedClient
+from shepherd_core import WebClient
 from shepherd_core import logger
 from shepherd_core.data_models import GpioTracing
 from shepherd_core.data_models.content import EnergyEnvironment
@@ -20,10 +20,10 @@ if __name__ == "__main__":
         path_task = path_here / "content/"
         path_fw = path_here / "content/fw/nes_lab/"
 
-    tb_client = TestbedClient()
     do_connect = False
     if do_connect:
-        tb_client.connect()
+        # connected -> publicly available data is queried online
+        WebClient()
 
     if not path_fw.exists():
         path_fw.mkdir(parents=True)
