+++ shepherd_core/examples/vsource_simulation.py
@@ -16,7 +16,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from shepherd_core import CalibrationEmulator
-from shepherd_core import TestbedClient
+from shepherd_core import WebClient
 from shepherd_core.data_models import VirtualSourceConfig
 from shepherd_core.vsource import VirtualSourceModel
 
@@ -38,16 +38,14 @@ src_list = [
 I_mcu_sleep_A = 200e-9
 I_mcu_active_A = 1e-3
 
-# for online-queries the lib can be connected to the testbed-server
+# For online-queries the lib can be connected to the testbed-server.
 # NOTE: there are 3 states:
-#    - unconnected -> demo-fixture is queried (locally)
-#    - connected -> publicly available data is queried online
-#    - logged in with token -> also private data is queried online
-tb_client = TestbedClient()
+# - unconnected -> demo-fixtures are queried (locally)
+# - connected -> publicly available data is queried online
+# - logged in with valid token -> also private data is queried online
 do_connect = False
-
 if do_connect:
-    tb_client.connect()
+    WebClient()
 
 for vs_name, v_hrv_mV, samples in product(src_list, v_hrv_mV_list, sample_dur_list):
     # prepare simulation
